Add files via upload
Project 4 Clustering

Customers who come for shopping in a mall can be grouped in a number of ways. Shopping pattern, spending pattern, shopping score, salary etc are some of the factors that can be used to divide the customers into different categories. The reason for these clusters is to help identify those customers who would be interested in certain products, offers and services.

Problem statement To create Customer cluster to identify the different types of customers